Dev docker private

+++ build_scripts/docker/dockerhub-build.sh
@@ -16,20 +16,22 @@ popd >/dev/null
 
 for app in ${O2_APPS[@]} ; do
    echo "Building ${app} docker image"
-   pushd ${app}
-   getImageName ${app} ${TAG}
-   docker rmi ${imagename}
-   docker build -t ${imagename} .
-   if [ $? -ne 0 ]; then
-     echo; echo "ERROR: Building container ${app} with tag ${TAG}"
+   if [ "${app}" ne "os-db" ] ; then
+     pushd ${app}
+     getImageName ${app} ${TAG}
+     docker rmi ${imagename}
+     docker build -t ${imagename} .
+     if [ $? -ne 0 ]; then
+       echo; echo "ERROR: Building container ${app} with tag ${TAG}"
+       popd
+       exit 1
+     fi
+     docker push $imagename
+     if [ $? -ne 0 ]; then
+       echo; echo "ERROR: Pushing container ${app} with tag ${TAG}"
+       popd
+       exit 1
+     fi
      popd
-     exit 1
    fi
-   docker push $imagename
-   if [ $? -ne 0 ]; then
-     echo; echo "ERROR: Pushing container ${app} with tag ${TAG}"
-     popd
-     exit 1
-   fi
-   popd
 done
